refactor: remove zero_agent_sessions table and migrate reads to chat_messages
Drop the legacy zero_agent_sessions table (id + chat_messages JSONB)
which has had no write path since the chat_messages table migration
(a5ee0da96). All remaining read paths are removed or rewritten:

- Delete dead routes: GET /api/agent/sessions (list, no callers),
  GET /api/zero/sessions/[id] (no frontend callers)
- Delete contracts: sessionsContract (list), zeroSessionsByIdContract
- Remove chatMessages field from SessionResponse (CLI only uses
  secretNames)
- Rewrite GET /api/agent/sessions/[id] to query agent_sessions directly
- Rewrite export-service to query agent_sessions + conversations for
  CLI history export instead of going through the JSONB column
- Delete zero-session-service.ts, zero-agent-session.ts schema
- Clean up test seeders and assertions

Closes #9588
